requirements.txt
pandas
openai
azure-ai-inference
azure-ai-resources
elasticsearch
streamlit
python-dotenv